check whether the listener belongs to a dialog
This is so that we don't mistake a Listener that's listening for a Dialog's packet, with a Listener that's listening for a dispatchable event (any other event that isn't dialog-related).

We're checking for that when the Listener is being registered, so that the Dialog can set the _is_dialog_listener attribute to True.